Improve reentrancy for deletes
